// src/app/models/listing.model.ts
export interface listing {
    address_line_1: string;
    address_line_2: string;
    city: string;
    contact_email: string;
    contact_name: string;
    contact_phone_number: string;
    description: string;
    number_of_bathrooms: string;
    number_of_bedrooms: string;
    price_per_month: string;
    state: string;
    zip_code: string;
    docId: string;
}